RF: moved helper tests.
